refactor(store): align listener isolation with the core precedent
type the internal callback as returning unknown instead of casting the result, drop the trailing return undefined and the multi-line comment, move the manager unit test beside the module, and let vitest own unhandled rejection detection instead of a manual process listener.
